change float to double

+++ oneflow/core/autograd/gradient_funcs/upsample.cpp
@@ -22,8 +22,8 @@ namespace one {
 
 struct UpsampleCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   float align_corners;
   std::string data_format;
   std::string interpolation;
@@ -54,8 +54,8 @@ Maybe<void> Upsample::Capture(UpsampleCaptureState* ctx, const TensorTuple& inpu
   ctx->requires_grad = inputs.at(0)->requires_grad();
   if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
   ComposedAttrMap composed_attrs(attrs, base_attrs_);
-  ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-  ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+  ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+  ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
   ctx->align_corners = JUST(composed_attrs.GetAttr<bool>("align_corners"));
   ctx->data_format = JUST(composed_attrs.GetAttr<std::string>("data_format"));
   ctx->interpolation = JUST(composed_attrs.GetAttr<std::string>("interpolation"));
@@ -80,8 +80,8 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample", Upsample);
 
 struct UpsampleNearest2DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   std::vector<int64_t> output_size;
   std::string data_format;
 };
@@ -97,8 +97,8 @@ class UpsampleNearest2D : public OpExprGradFunction<UpsampleNearest2DCaptureStat
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-    ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+    ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+    ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
     }
@@ -129,8 +129,8 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_nearest_2d", UpsampleNearest2D);
 
 struct UpsampleBilinear2DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   bool align_corners;
   std::vector<int64_t> output_size;
   std::string data_format;
@@ -147,8 +147,8 @@ class UpsampleBilinear2D : public OpExprGradFunction<UpsampleBilinear2DCaptureSt
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-    ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+    ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+    ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
     ctx->align_corners = JUST(composed_attrs.GetAttr<bool>("align_corners"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
@@ -180,7 +180,7 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_bilinear_2d", UpsampleBilinear2D);
 
 struct UpsampleLinear1DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float scale_factor;
+  double scale_factor;
   bool align_corners;
   std::vector<int64_t> output_size;
   std::string data_format;
@@ -197,7 +197,7 @@ class UpsampleLinear1D : public OpExprGradFunction<UpsampleLinear1DCaptureState>
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->scale_factor = JUST(composed_attrs.GetAttr<float>("scale_factor"));
+    ctx->scale_factor = JUST(composed_attrs.GetAttr<double>("scale_factor"));
     ctx->align_corners = JUST(composed_attrs.GetAttr<bool>("align_corners"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
@@ -229,7 +229,7 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_linear_1d", UpsampleLinear1D);
 
 struct UpsampleNearest1DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float scale_factor;
+  double scale_factor;
   std::vector<int64_t> output_size;
   std::string data_format;
 };
@@ -245,7 +245,7 @@ class UpsampleNearest1D : public OpExprGradFunction<UpsampleNearest1DCaptureStat
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->scale_factor = JUST(composed_attrs.GetAttr<float>("scale_factor"));
+    ctx->scale_factor = JUST(composed_attrs.GetAttr<double>("scale_factor"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
     }
@@ -275,8 +275,8 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_nearest_1d", UpsampleNearest1D);
 
 struct UpsampleBicubic2DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   bool align_corners;
   std::vector<int64_t> output_size;
   std::string data_format;
@@ -293,8 +293,8 @@ class UpsampleBicubic2D : public OpExprGradFunction<UpsampleBicubic2DCaptureStat
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-    ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+    ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+    ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
     ctx->align_corners = JUST(composed_attrs.GetAttr<bool>("align_corners"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
@@ -326,8 +326,8 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_bicubic_2d", UpsampleBicubic2D);
 struct UpsampleNearest3DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
   float depth_scale;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   std::vector<int64_t> output_size;
   std::string data_format;
 };
@@ -343,9 +343,9 @@ class UpsampleNearest3D : public OpExprGradFunction<UpsampleNearest3DCaptureStat
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->depth_scale = JUST(composed_attrs.GetAttr<float>("depth_scale"));
-    ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-    ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+    ctx->depth_scale = JUST(composed_attrs.GetAttr<double>("depth_scale"));
+    ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+    ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
     }
@@ -377,8 +377,8 @@ REGISTER_OP_EXPR_GRAD_FUNCTION("upsample_nearest_3d", UpsampleNearest3D);
 struct UpsampleTrilinear3DCaptureState : public AutoGradCaptureState {
   bool requires_grad;
   float depth_scale;
-  float height_scale;
-  float width_scale;
+  double height_scale;
+  double width_scale;
   bool align_corners;
   std::vector<int64_t> output_size;
   std::string data_format;
@@ -395,9 +395,9 @@ class UpsampleTrilinear3D : public OpExprGradFunction<UpsampleTrilinear3DCapture
     ctx->requires_grad = inputs.at(0)->requires_grad();
     if (!ctx->requires_grad) { return Maybe<void>::Ok(); }
     ComposedAttrMap composed_attrs(attrs, base_attrs_);
-    ctx->depth_scale = JUST(composed_attrs.GetAttr<float>("depth_scale"));
-    ctx->height_scale = JUST(composed_attrs.GetAttr<float>("height_scale"));
-    ctx->width_scale = JUST(composed_attrs.GetAttr<float>("width_scale"));
+    ctx->depth_scale = JUST(composed_attrs.GetAttr<double>("depth_scale"));
+    ctx->height_scale = JUST(composed_attrs.GetAttr<double>("height_scale"));
+    ctx->width_scale = JUST(composed_attrs.GetAttr<double>("width_scale"));
     ctx->align_corners = JUST(composed_attrs.GetAttr<bool>("align_corners"));
     if (base_attrs_.find("output_size") != base_attrs_.end()) {
       ctx->output_size = JUST(composed_attrs.GetAttr<std::vector<int64_t>>("output_size"));
